+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ## unreleased
 
+- allow to specify included and excluded schemas
 - fixed an issue which caused identity upon table creation to be normal columns
 - order schemas in diffs by first occurrence of a relation to solve dependency issues
 - fixed rule creation and drop statements